requirements.txt


PyQt6==6.7.1
PyQt6-3D==6.7.0
PyQt6-Charts==6.7.0
PyQt6-WebEngine==6.7.0
pymavlink>=2.4.47
pyserial==3.5
arcgis>=2.4.1
folium>=0.17.0
requests>=2.32.0
numpy>=1.24.0
pandas>=2.0.0
geopy>=2.4.0
opencv-python>=4.9.0
